Directly export Formio
The parameters for the exported function that returns the Formio class is unnecessary. The parameters are no longer used (except _baseUrl which can be set with Formio.setBaseUrl()). I removed it so that it would be easier for users to get access to the Formio class.

Also renamed the standalone export name from `formiojs` to `Formio` to better reflect this change.