Fix bugs with Select and Checkbox component
FOR-1646:
Because of tricky focus redirection to focusableElement, Choices onBlur
handler was triggered and this behavior forced dropdown to be closed.
Now we add this focus redirection only when search is enabled.

FOR-1647:
Because false was considered as empty value for Checkbox component,
all newly created checkboxes was set to true due to restoreValue
method. Now undefined is considered as empty value for Checkbox.