Enable Custom Builder Components Pre-Defined Key
The Business need this fix pertains to is to allow the Custom Formbuilder to pre-define Key names. Although the component could have its key pre-defined, it would always be overwritten by the "Ensure this component has a key" logic due to the fact component.keyModified is undefined.

The proposed fix will now check if the new component has a predefined key, and if so it will leave the key alone. Should no key be predefined, it will simply continue its regular logic of defining one for the component.