Fix form embedding in async DOM (replace document.write() by DOM insert).
When a document is asynchronously rendered, a document.write() from a
different source (eg a script) is ignored and results into following
warning. Also the embedding of the form fails with an error `Uncaught
TypeError: p is null`.

Warning: A call to document.write() from an asynchronously-loaded external
script was ignored.